Split common sections into links and updates
Break down by feature, not everything that has updates will also
require metadata, for instance.

* Move `part_of` into linkable module so that all part_of uses look at
the same fields
